Transfers: Add Dynamic TransferWaitTime SourceRankingStrategy
We add a dynamic TransferWaitTime SourceRankingStrategy to estimate for each source node, how long a new transfer has to wait to begin by approximating how long it will take to process the current queue. Source nodes with smaller estimated wait times are preferred. We introduce randomization so that we are able to distribute requests across multiple sources if their estimated wait times are similar and so that even sources that we think have the longest wait time will still get picked a small amount of the time. We choose below average wait sources 90% of the time and above average wait sources 10% of the time, and proportionally pick randomly based on wait time.